- Updated license headers for year 2021 - Upgraded parent to 8.0.0 -> Upgrade to Java 11 (Java 8 no longer supported) - Maven plugin findbugs (obsolete) replaced with spotbugs

+++ src/main/java/org/ow2/authzforce/core/pdp/api/policy/PolicyVersion.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2012-2020 THALES.
+ * Copyright 2012-2021 THALES.
  *
  * This file is part of AuthzForce CE.
  *
@@ -92,7 +92,7 @@ public final class PolicyVersion implements Comparable<PolicyVersion>
 				throw new IllegalArgumentException("Invalid Policy(Set) Version: '" + version + "'", e);
 			}
 
-			if (number.intValue() < 0)
+			if (number < 0)
 			{
 				throw new IllegalArgumentException("Invalid Policy(Set) Version: '" + version + "'. Number #" + i
 						+ " (=" + number + ") is not a positive integer");
